+++ tests/qiskit_backend_test.py
@@ -18,12 +18,12 @@ from typing import Optional, Any
 import numpy as np
 import pytest
 
-from qiskit import QuantumCircuit, execute  # type: ignore
+from qiskit import QuantumCircuit  # type: ignore
 from qiskit.primitives import BackendSampler  # type: ignore
 from qiskit.providers import JobStatus  # type: ignore
 from qiskit_algorithms import Grover, AmplificationProblem, AlgorithmError  # type: ignore
 from qiskit_aer import Aer  # type: ignore
-
+from qiskit_ibm_runtime import QiskitRuntimeService, Session, Sampler  # type: ignore
 from qiskit_ibm_provider import IBMProvider  # type: ignore
 
 from pytket.extensions.qiskit import (
@@ -66,7 +66,7 @@ def test_samples() -> None:
     b = AerBackend()
     for comp in (None, b.default_compilation_pass()):
         tb = TketBackend(b, comp)
-        job = execute(qc, tb, shots=100, memory=True)
+        job = tb.run(qc, shots=100, memory=True)
         shots = job.result().get_memory()
         assert all(((r[0] == "1" and r[1] == r[2]) for r in shots))
         counts = job.result().get_counts()
@@ -78,12 +78,12 @@ def test_state() -> None:
     b = AerStateBackend()
     for comp in (None, b.default_compilation_pass()):
         tb = TketBackend(b, comp)
-        job = execute(qc, tb)
+        job = tb.run(qc)
         state = job.result().get_statevector()
         qb = Aer.get_backend("aer_simulator_statevector")
         qc1 = qc.copy()
         qc1.save_state()
-        job2 = execute(qc1, qb)
+        job2 = qb.run(qc1)
         state2 = job2.result().get_statevector()
         assert np.allclose(state, state2)
 
@@ -93,12 +93,12 @@ def test_unitary() -> None:
     b = AerUnitaryBackend()
     for comp in (None, b.default_compilation_pass()):
         tb = TketBackend(b, comp)
-        job = execute(qc, tb)
+        job = tb.run(qc)
         u = job.result().get_unitary()
         qb = Aer.get_backend("aer_simulator_unitary")
         qc1 = qc.copy()
         qc1.save_unitary()
-        job2 = execute(qc1, qb)
+        job2 = qb.run(qc1)
         u2 = job2.result().get_unitary()
         assert np.allclose(u, u2)
 
@@ -107,26 +107,50 @@ def test_cancel() -> None:
     b = AerBackend()
     tb = TketBackend(b)
     qc = circuit_gen()
-    job = execute(qc, tb, shots=1024)
+    job = tb.run(qc, shots=1024)
     job.cancel()
     assert job.status() in [JobStatus.CANCELLED, JobStatus.DONE]
 
 
+# https://github.com/CQCL/pytket-qiskit/issues/272
+@pytest.mark.xfail(reason="Qiskit sampler not working")
 @pytest.mark.skipif(skip_remote_tests, reason=REASON)
-def test_qiskit_counts(nairobi_emulator_backend: IBMQEmulatorBackend) -> None:
+def test_qiskit_counts(ibmq_qasm_emulator_backend: IBMQEmulatorBackend) -> None:
     num_qubits = 2
     qc = QuantumCircuit(num_qubits)
     qc.h(0)
     qc.cx(0, 1)
     qc.measure_all()
 
-    s = BackendSampler(TketBackend(nairobi_emulator_backend))
+    s = BackendSampler(TketBackend(ibmq_qasm_emulator_backend))
 
     job = s.run([qc], shots=10)
     res = job.result()
 
     assert res.metadata[0]["shots"] == 10
     assert all(n in range(4) for n in res.quasi_dists[0].keys())
+
+
+# https://github.com/CQCL/pytket-qiskit/issues/272
+@pytest.mark.xfail(reason="Qiskit sampler not working")
+@pytest.mark.skipif(skip_remote_tests, reason=REASON)
+def test_qiskit_counts_0() -> None:
+    num_qubits = 32
+    qc = QuantumCircuit(num_qubits)
+    qc.h(0)
+    qc.cx(0, 1)
+    qc.measure_all()
+
+    _service = QiskitRuntimeService(
+        channel="ibm_quantum",
+        instance="ibm-q/open/main",
+        token=os.getenv("PYTKET_REMOTE_QISKIT_TOKEN"),
+    )
+    _session = Session(service=_service, backend="ibmq_qasm_simulator")
+
+    sampler = Sampler(session=_session)
+    job = sampler.run(circuits=qc)
+    job.result()
 
 
 def test_architectures() -> None:
@@ -137,7 +161,7 @@ def test_architectures() -> None:
         # without architecture
         b = MockShotBackend(arch=arch)
         tb = TketBackend(b, b.default_compilation_pass())
-        job = execute(qc, tb, shots=100, memory=True)
+        job = tb.run(qc, shots=100, memory=True)
         shots = job.result().get_memory()
         assert all(((r[0] == "1" and r[1] == r[2]) for r in shots))
         counts = job.result().get_counts()